requirements.txt
pygame
pyserial
tk